// heatcheck-app/worker/types.ts
export interface ParkInfo {
  venueId: number;
  name: string;
  city: string;
  lat: number;
  lon: number;
  hrFactor: number;
  kFactor: number;
  cfBearingDegrees: number;
  isDome: boolean;
}